requirement.txt
sklearn
imageio
click
tensorflow-gpu == 1.13.0
keras == 2.2.4
